requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
handler==1.3.0
idna==3.4
Pillow==9.2.0
pony==0.7.16
psycopg2==2.9.4
requests==2.28.1
urllib3==1.26.12
vk-api==11.9.9
